+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+
+- `opentelemetry-instrumentation-pika` Instrumentation for `channel.consume()` function 
+  ([#2397](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/2397)))
 
 ### Breaking changes